requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
billiard==4.2.1
celery==5.5.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
concurrent-log-handler==0.9.26
cron-descriptor==1.4.5
Django==5.2
django-celery-beat==2.8.1
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
drf-standardized-errors==0.14.1
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kafka-python==2.2.8
kombu==5.5.3
portalocker==3.1.1
prompt_toolkit==3.0.51
psycopg2==2.9.10
PyJWT==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
redis==6.1.0
referencing==0.36.2
rpds-py==0.25.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
